ginstall: Add --once option
ginstall will now check for the completion file when given the "--once"
option.

The reason for this is a race condition that abf@ observed on GFMN. If
cwmp exits due to OOM killer intervening, and ginstall instance 1 is
running, then cwmp attempts to run ginstall instance 2, often times
instance 2 takes so long to get to the portion of code that attempts to
acquire a file lock that instance 1 will run to completion, and instance
2 will run through normally without ever checking for the completion
file left behind by instance 1.

Change-Id: I9a53e8d4cd40f5af3e6a296713a559f44aa6e129