gencert: add a new gfiber-device-cert debian package.
It was kind of a pain to generate device certs for newly installed
devices.  This package auto-generates a key and self-signed certificate when
it's installed (or whenever you run "gfiber-device-cert -f") and prints
the public cert to stdout when you later run "gfiber-device-cert".  You
can then add the cert to your favourite service to allow this machine
to authenticate.

Also modify the gfiber-loguploader package to depend on this package, since
it's one of the things that uses the device cert.

(The basic key generation feature is a much-reduced version of the
manufacturing certificate script from dgentry.)

Change-Id: I02f56c059af0b3075143e9afb34997a0da3e8ccb