/bin/wifi: add an extra level of -X, and add -Y (disable rekey)
Now -XX is the equivalent of the old -X.  With just one -X, we enable peer
rekeying, which was disabled by default, in case that improves security at
all.

With -Y, we disable *all* rekeying.  This may reduce security, but if you
suspect a bug in rekeying, this can help narrow it down until we fix the
bug.

Change-Id: I7f846c6fa894ba828462ec55a8043fd76e74fc60