/bin/wifi:  Quantenna interface in bridge iff AP mode.
Remove the Quantenna driver interface from the bridge in client mode,
so that we don't bridge open networks with the user's LAN during
provisioning.  This also has the benefit of making the local network
topology more similar to CPE2, which has separate wcli* client
interfaces which are never in the bridge.

Add the Quantenna driver back into the bridge in AP mode.

Change-Id: I297a2e7f267400af974cac4f4b24b8506afaba62