bin/wifi: avoid channel conflicts starting APs.
hostapd will fail to start with an uninformative error, "Failed to set
beacon parameters," if you try to start a VAP with it and there is an
existing AP configured on a different channel that uses the same radio.

Work around this by letting the first channel setting win. A later CL
will provide a way for us to actually change channel in these
circumstances.

Change-Id: I9a5e183bb3e4ab4c96c132d8cbe73ffa78715ef6