conman:  Enable/disable ACS autoprovisioning.
After adding a wired route, enable ACS autoprovisioning.  Before
removing it, delete ACS autoprovisioning.  This is used in
go/fibercl/70020.

There is a minor race condition without a corresponding change in
apman:  If the wired and wireless connections are both up, and then
conman dies, the deathrattled apman could remove the wired route,
leaving the file which enables ACS autoprovisioning in place.

Change-Id: I5dcd2fae84f99b578a964bd0c025ce2bd5ee1699