platform: add libexperiments
This provides a library for C/C++ programs that enables them to easily
register experiments and check if they are enabled.

In order to accommodate programs like Sagesrv which may be checking
experiment states at higher frequencies, a caching logic is used that
stores the current state of registered experiments and updates them in
(configurable) intervals (e.g., Sagesrv defaults to 60 secs).

Please see experiments.h for details on usage.

Change-Id: I2d55161960dc5010beb666324666f6fd1f210f97