/bin/wifi:  Update provisioning vendor IE feature IDs.
We don't need to separate these by band if the provisioning network AP
includes the IE in its management frames, rather than the same
device's WLAN AP doing so.  We also don't need separate flags for them
at all this way, since we already have the ssid from --ssid, and can
enable the IE whenever --hidden_mode is enabled.

Also, Jetstream got ID 2 while the initial implementation of this was
in flight; we reserve 3 in cl/120365499.

Change-Id: I22794ecde9c3a8ee4d8cfb2a2b8da3b1a5bea5e6