conman:  Fix a flaky test.
Previously, there was code that would run conman until it updated its
cached interface state, and then until it scanned.  But occasionally
the scan would happen before or during the main loop iteration which
updated the interfaces, in which case run_until_scan would never
terminate.

This fixes the issue by adding a run_until_* function which combines
the two.

Change-Id: I520a6ae1fb65cdfa5ff8a865a232d8cfaf3112af