conman:  Tweak FrenzyWPACtrl.attached.
A normal WPACtrl can attach to wpa_supplicant even when not connected
to a network.  FrenzyWPACtrl cannot emulate this perfectly, but
currently (by requiring that there is an SSID) its `attached` method
requires that wpa_supplicant is actually connected to a network.
Removing this requirement and only requiring that the Quantenna chip
is in client mode is closer to the real WPACtrl behavior.

Change-Id: Ic5366baad7b784be2576b6df987ed0614131797b