conman:  Deflake test.
The test contained a race condition which could result in a scan
happening before expected, and then an infinite loop waiting for a
scan that never happened.  The fix is easy (and obvious in
retrospect):  don't enable the scan results until immediately before
the run_until_interface_update_and_scan call.

Change-Id: I8fbc8e7377cebfa8f4b471ec9806da50e2cd2e37