asustax: print genus and species.
genus is intended to be displayed to the user. Make sure it
contains the string "ASUS" so as to be a bit more meaningful.

species is intended to be the most specific identifier we have.
Print the model name we get from the ASUS Discovery Packet.

Replace BUILD_ASUS & BUILD_WIFI_UTILS with BUILD_LIBNL_UTILS
as that is the salient thing needed.

Change-Id: Ib35647abde85c077fab7baeca00e6da030090c5b