dhcpvendortax: genus & species support.
We've converged on a format of:
   taxonomy: type label Short Name;Exact Model
in the other taxonomy mechanisms. Implement that support in
dhcpvendortax as well, which was really only outputting the
short name.

This involved converting the substring matchers to regex, so
they could extract model information from the string. That in
turn means that the regex cases previously handled in check_specials
can be lumped in with the other regexes now.

Add a test case for every regexp.

Also convert the gperf file to use | as its delimeter, as the
other taxonomic gperf files now do.

Change-Id: I9596bbfda030d983cc5bc41de03594d9aa3d7f9c