gfch100: Fix for presterastats, handle cpss LUA CLI being left in a context.
This can happen, for example, if someone is manually using cpss_cmd and
leaves the CLI in 'debug' context. If this happens, the "show
interfaces..." command from presterastats fails because it needs to run
in the 'exec' context. Prefixing an exec mode command with "do ..." lets
it run without leaving the other context.

Change-Id: I6962b6c60d88a7f61f111ce08a3ee11f63ffded6