ginstall: add flash unlock support for gflt
sysvar and GFLT110 have flash locking behavior that can cause ginstall
to fail to install until the next reboot (in the new uboot we turn off
flash protection before boot) however flash can be locked from Linux.

If this is a FiberJack, make ginstall unlock the mtd partition we are
writing to using the flash_unlock tool.

Change-Id: I9ec8b3a04261e216f52f90632a0c2745f3471316