/bin/wifi:  setclient performance improvements.
Use wpa_cli reconfigure rather than stopping and starting
wpa_supplicant.  This makes associating with a new AP much faster.
The total /bin/wifi time to reassociate is now ~3 seconds, down from
10-15.  Most of this is /bin/wifi itself, but reassociating can still
take ~1s.

If an AP is running as well, it will still take time to stop and
restart that, but once b/30140131 is resolved we may be able to remove
that as well.

Change-Id: Id26f7ab9b3c82aba9c06f7889c33d13455d5b5eb