platform: change ipv6 multicast address for stats
The current IPv6 multicast address used to send and receive stats from
the ONU to the RG is an invalid multicast address and seems to have
stopped working. This commit creates (and documents) the change to the
new IPv6 multicast address.

Format of an IPv6 multicast address:

+--------+----+----+----------------------------------------+
|    8   |   4   |   4   |            112 bits              |
+--------+----+----+----------------------------------------+
|11111111| flags | scope |            group ID              |
+--------+----+----+----------------------------------------+

IPv6 multicast addresses are all addresses inside the prefix ff00::/8.
There are four bits each for flags and scope. The final 112 bits
specify the "group id".

Of the four "flag" bits (defined in RFC 3513) only the low-order bit
is currently used. It indicates that an address is 'transient' or
'non-permanent'. Permanently assigned addresses (transient bit is 0)
are deemed "well-known" and a formal list of addresses is maintained
by IANA.

The flags portion of the multicast group address for statcatcher
should set this low-order bit to 1.

The next 4 bits is the "scope" field, which is used to limit the scope
of the multicast group. Possible values are:

	0  reserved
	1  interface-local scope
	2  link-local scope
	3  reserved
	4  admin-local scope
	5  site-local scope
	8  organization-local scope
	E  global scope
	F  reserved

"Link-local" appears to be the most appropriate due to the constraint
that it must be "physically connected" which "admin-local" is not, as
defined in RFC 3513.

Change-Id: I89377c3293fb7691e9f364618d98ef3dd6b562b3