waveguide: make taxonomy combine genus + species.
Re-format the output format to retain the existing
format where the model is identified as the second tuple
in a semicolon-separated list. The third tuple is the
performance information.

The first tuple used to be the chipset, but the expected
uses for the chipset info never worked. We thought we would
be able to work around all kinds of chipset bugs, which we
thought would be rampant if only we knew how to look for them
It turns out that Wifi bugs are way more subtle than that.
Who knew?

Change-Id: I62d4414148053088b7ec07e9dfa79a48e6fab9ca