ssdptax: also send M-SEARCH
ssdptax was written to rely on MiniSSDPd for its list
of nearby SSDP devices, but that only finds a subset
of devices. MiniSSDPd tracks NOTIFY messages sent by
new devices, it never sends an M-SEARCH.

Some devices never send NOTIFY, they only respond to M-SEARCH.

However there appear to be many devices which do not
respond to M-SEARCH, they only send their own NOTIFY
messages periodically. MiniSSDPd is very useful for
those, because it is running all the time.

Supplement ssdptax to also send its own M-SEARCH, and
combine that with the list it gets from MiniSSDPd.
In testing, this retrieves information about the
maximal number of devices.

Change-Id: I139b1a023dfd98b035729c71e678af05cc61fb7b