conman:  Also prioritize open APs by OUI and RSSI.
Before this commit, open APs are only prioritized by presence of the
GFiber provisioning OUI.  Frenzy devices cannot see this, so OUI
serves as a proxy (it will be rare for an open GFiber AP not to
provide ACS access).  And further prioritizing by RSSI should be a
minor performance boost.

BUG=30376078

Change-Id: Ie59457e3f18fd097d1400ecd5e280e721459052f