conman:  Determine client connection status using key_mgmt.
We don't want to reassociate with new network credentials as long as
we're still happily connected.  We were determining whether we were
connected to the WLAN using SSID, but since that could be part of
changed credentials, use key_mgmt instead.  We only connect to the
WLAN and to open networks so this is sufficient.

BUG=30448048

Change-Id: I11c3bc7626423220642be882a4e596f4b7e2bac3