ledpattern: Initial implementation of LED tap codes
The GFLT's have a physical reset button that can be "tapped" or pressed
for less that a second. The LEDs will then pulse a specific diagnostic
code that can be looked up to to give a higher level of diagnostic
information than would normally be available in the field.

The LED tap codes are described in http://go/fiberjack-leds

Change-Id: I1a21505017f8e133edfd78f8a6d12c1d52cfcfab