wifi_files: test quoting and correct switch.
Add an explicit test for quote and backslash, which
are printable but must nonetheless be escaped in JSON.
The code does handle them properly already, but those
two characters have been problematic on the windcharger-45
branch so make sure they never get accidentally broken.

Also fix the break; in print_ssid_escaped's switch
statement. Functionally it did work (it is the last
case statement so it falls off the end) but its still
wrong so fix it.

Change-Id: Ifadbe81881d56b01aa0411ed233386941aed247c