conman:  Always update routes.
Previously, conman tried to avoid interacting with the routing table
if its connection status hadn't changed.  But this didn't quite work,
because the 'before' value might have been updated elsewhere since the
connection status changed, and so we might not detect a newly working
connection and add the appropriate routes.

This was a premature optimization, so rather than try to make it work
it has been removed.  The code which applies changes to the routing
table avoids trying to add routes which already exist anyway.

Change-Id: Ifb911b87e8e2440ed4680e10785547d33954ecf6