Implement the gpio-mailbox api for brcm w/ NEXUS.
This change adds an implementation of the low-level
BRCM API for gpio-mailbox that uses the BRCM provided
nexus API (via nxclient). In order to facilitate this
change, the original broadcom.c file is split into
the public pin API and a new private HAL API (defined
in brcm-platform.h).

The old mmap implementation is moved to brcm-direct.c
while the nexus implementation lives in brcm-nexus.c.
The backing implementation for a given system can be
swapped as necessary, although the nexus API only supports
the GFHD254 as of now.

Change-Id: I7488c7856179d3082b8cb15721d4d06c786c42ae