conman:  Fallback to wpa_cli when control interface fails.
For as yet unknown reasons, requests to the the wpa_supplicant control
interface sometimes fail.  When that happens, we can fail over to
using a subprocess call to the wpa_cli.

The unreliable code is left in place for two reasons:

- Recently added code may help debug it.
- It's slightly more efficient to avoid unnecessary subprocess calls.
- We want to use the control interface anyway in order to receive
  events.  (It's unclear whether the request bugs also affect this,
  which is further motivation to debug that problem.)

Seperately, the Frenzy test was actually previously incorrect
(simulating the request failure, which could not occur on Frenzy
because we don't use wpactrl on Frenzy), and has been fixed.

BUG=30638247

Change-Id: Ic5d9bf597129810a3690a0bd2bffdebb4dd4d59b