conman:  Don't interrupt provisioning.
The WLAN retry period was much too low (15s), and this was
interrupting provisioning (which might be needed to get an updated
config).  The default period has been increased to 120s, and
provisioning won't be interrupted to retry it.

Also added unit tests for this case and for reprovisioning in general,
and slightly tightened up tests in general.

Change-Id: I8049679964974c068fb04cf384918a8bcbe35e71