conman:  Stop existing APs at startup with no wired connection.
If an AP was already running when conman starts (e.g. from the 'wifi
restore' call in S50wifi), and Catawampus says to run an access point,
conman would leave it running - even with no wired connection.

Now conman also checks for a working wired connection in order to
leave the AP up.

Change-Id: I00e0561971b15970c28d92dc17794b8623e3b72f