conman: always initialize WPA control interface on startup.
Previously, conman only initialized the WPA control interface if the
client interface was not already up. In the case of a provisioned
device, the client interface will be started at boot by "wifi restore".
As a result, conman does not initialize the WPA control interface. When
processing the command file, conman will incorrectly think the client
interface is down, clear the routes, and run "wifi setclient". Because
the WiFi configuration has not changed, "wifi setclient" is a no-op and
dhclient is not restarted. This results in a state where the client
interface is associated but there are no routes since no new DHCP leases
were received.

Change-Id: I8c66d626e9566d1526cca018332d44489ec66a1e