/bin/wifi: restart dhclient on wpa_supplicant reconfigure.
Commit 1b1a151de35a9f06ecf0b7311499ec365da36cf2 introduced an
optimization to avoid restarting wpa_supplicant if it is already running
by sending it a "reconfigure" command. This caused it to bypass the
ifplugd action that starts dhclient. As a result, no new DHCP lease is
requested when switching APs (such as from a provisioning AP to a
restricted AP or to another provisioning AP).

Change-Id: Ic0fb633e7b275a6292c681126eee02b15de80849