Disable alivemonitor for hostapd on windcharger.
Windcharger does not currently use waveguide, which generates important
events that keep the hostapd alive file updated. Without waveguide,
the other events that keep the alive file updated are generated externally
such as beacons and probes, and are less available in some installations
leading to hostapd getting killed when it is operating normally.

Also add an experiment to disable alivemonitor on hostapd.

see b/32376077

Change-Id: Ic1a4d593035a595fab6bc385fcb2ec0bb3f733e5