conman:  Don't use wpa_supplicant control interface.
The control interface is unreliable, both for requests and events.
Stop using it and go with the slightly more expensive but much simpler
approach of always talking to wpa_cli, which so far seems much more
reliable.

While writing and debugging this, it was useful to improve the
printing of mocked access points.

Change-Id: I05205abda499646aa35cd63bbc5f68fd331b34c1