Set the aon gpios to be outputs for the leds.
Adjust the pwm frequency to the pwm feeding the power and change
the gpios that control the on/off to be outputs only.  The gpios that
connect to the switches are connected to the LDK controller but we
get better control with the the PWM.  We left the LDK working so older
evt boards would continue to work (the power pwm was added around evt3).
However we need to remove the LDK code in order to get full brightness
on the leds.

Change-Id: I0cdea98499c842e2902e8075038f4e7232211eb9