/bin/wifi:  Use new AP's channel for existing APs.
When specifying an AP channel on a phy which already has another AP
running on another interface (e.g. a _portal interface), restart the
existing AP with the new channel before starting the new AP.

This is a partial reversion of commit 62f7841, which just used the
existing AP's channel for the new AP.  This was a problem because with
two APs running on the same phy, it was impossible to change the
channel of either.

BUG=31074359

Change-Id: Ie6d5fff29cb77e08a18504bfc64fc7662f77f310