taxonomy: qualify Amazon Dash Button with OUI.
The OS signature for the Dash Button is very, very simple and
does not differentiate other embedded devices like Honeywell
thermostats which use the same Wifi chipset and have the same
DHCP signature.

Qualify the Dash Button with the OUI instead of the OS signature.
We'll leave the OS signature in dhcp.py as a record that it has
aliases.

Fixes b/34062262

Change-Id: Ia84cd42fd281022603aca2263648f46ce0a9c96f