Make `wifi setclient` fast again.
PyCrypto's PBKDF2 implementation does a lot of work in Python. This is
OK on a powerful workstation, but not OK (it takes 8 seconds!) on a TV
box.

Use OpenSSL's PBKDF2 implementation instead, via ctypes. This takes 0.25
seconds in a microbenchmark, of which ~half is the time to load the
Python interpreter.

Change-Id: I7ae4805f4aa9971d59c495037aa8b7a3e61352d3